Removed the main page Moved the main page to the catalog page
