--- src/pages/dataCatalog/__tests__/DataCatalog.test.tsx
@@ -1,9 +1,0 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
-import DataCatalog from '../DataCatalog';
-
-it('renders DataCatalog page without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(<DataCatalog />, div);
-  ReactDOM.unmountComponentAtNode(div);
-});
