+++ src/saga/index.ts
@@ -1,6 +1,6 @@
 import { all, call } from 'redux-saga/effects';
 import { configSaga } from '../config/configSaga';
-import { dataSaga } from '../pages/mainPage/saga';
+import { dataSaga } from '../pages/dataCatalog/saga';
 
 export function* rootSaga() {
   yield all([call(configSaga), call(dataSaga)]);
